fix typo found by @cameronelliott

+++ RRL/Round1/xdp_rrl_per_ip.c
@@ -21,7 +21,7 @@
 typedef __u8 uint8_t;
 typedef __u16 uint16_t;
 typedef __u32 uint32_t;
-typedef __u32 uint64_t;
+typedef __u64 uint64_t;
 #define memcpy __builtin_memcpy
 
 /*
